map renders and zooms
